backend/pyproject.toml
[project]
name = "mis-eventos"
version = "0.1.0"
description = ""
authors = [
    {name = "Diego Marin",email = "devdmarin@gmail.com"}
]
license = {text = ""}
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.133.0,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (==1.7.4)",
    "alembic (>=1.18.4,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "redis (>=7.2.0,<8.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==4.1.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
